Initial drop of 'regex parser'
This is a syntax parser for a human readable regex language.
The parser returns an abstract syntax tree. The module describes
the regex language. It is very basic, and will no doubt grow over time.

Right now this code has some basic unit tests but isn't hooked up to
anything. The idea is that this code will be added to the REPL, and used
to interact with the regex module.