Add 'kind' field to Vertex subclasses
Use this from now on to determine the type of a Vertex subclass.

Remove use of isinstance in filter methods. We were using isinstance
to determine the type of a vertex. However, profiling determined that
this was insanely costly.

E.g. on a 55,000 node 205,000 edge graph calls to 'isinstance' were in
the top 10 of time spent for the application as a whole (due to their
use in filter methods). The isinstance itself was taking 4s and the
abcmeta's __isinstance__ was taking up 14s!