Remove Filter and FilterSet classes
FilterSet wasn't used anymore.

Filter was holding kind of weird state (the calculated linear operator
applied to a vector). This now exists as 'linop_calculate_values()'
as part of the Regex class. This I think is much cleaner.

Filter also had the graph and number of nodes. Small refactors that pass
this dependency around. Added a cache of the number of nodes to the
context. Admiitedly, this is kind of weird and probably shouldn't be
here.

Removed bin/vinge_regex. It had no functionality that doesn't now exist
in bin/vinge.